requirements.txt
aiohttp>=3.9.0
web3>=6.15.0
eth-account>=0.11.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
websockets>=12.0
